requirements.txt
amqp==5.1.1
aniso8601==9.0.1
arrow==1.2.3
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.5
cycler==0.11.0
dateparser==1.1.1
Flask==2.2.2
Flask-JWT==0.3.2
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.0
fonttools==4.37.4
greenlet==1.1.3
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib==3.6.1
moment==0.12.1
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.31
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
redis==4.4.0
regex==2022.3.2
six==1.16.0
SQLAlchemy==1.4.41
times==0.7
tzdata==2022.5
tzlocal==4.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.9.0
